Refactor side methods (#27)
* Consolidate extract_values_from_output into CheckType.get_value

* Simplify evaluator's file to contain only entrypoints

* Small changes to improve comprehension

* Small code refactor and comments

* rename check_type name and include the main class in package init for easier import

* Consolidate utils functions into less files, with some purpose
